requirements.txt
torch==1.13.1
torchvision==0.14.1
numpy==1.24.1
scikit-learn==1.2.0
scipy==1.10.0
pre-commit==2.21.0
Pillow==9.4.0
pandas==1.5.2
matplotlib==3.6.3